requirements.txt
gitpython
jira
termcolor
lxml
bs4
python-gitlab
